Fix incomplete instance data returned after build failure
This change fixes a race in _cleanup_build_artifacts. We were updating
the instance mapping to point at the cell in which the instance was
created before the instance record was complete, i.e. before the related
BDMs and tags were created in the cell DB. Updating the instance mapping
exposes the cell's version of the instance to the API. If the API happened
to fetch it before it was finished being created it would return
incomplete data.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Closes-Bug: #1820337
Change-Id: If966eb1161c842ff49aa530e4482dbca87b61a3e