Add functional regression test for bug 1669054
Change Ic3968721d257a167f3f946e5387cd227a7eeec6c in Newton
started setting the RequestSpec.ignore_hosts field to the
source instance.host during resize/cold migrate if
allow_resize_to_same_host=False in config, which it is by
default.

Change I8abdf58a6537dd5e15a012ea37a7b48abd726579 also in
Newton persists changes to the RequestSpec in conductor
in order to save the RequestSpec.flavor for the new flavor.
This inadvertently persists the ignore_hosts field as well.

Later if you try to evacuate or unshelve the server it will ignore
the original source host because of the persisted ignore_hosts value.
This is obviously a problem in a small deployment with only a few
compute nodes (like an edge deployment). As a result, an evacuation
can fail if the only available host is the one being ignored.

This adds a functional regression recreate test for the bug.

Change-Id: I6ce2d6b1baf47796f867aede1acf292ec9739d6d
Related-Bug: #1669054