Add functional test for the JsonFilter
Before this we had no functional testing of the JsonFilter.
Given it is a pretty complicated filter we should have
functional testing for it, so this adds a simple test which
is using a commonly reported pattern which is filtering
hosts based on hypervisor_hostname for trying to build
a baremetal server on a specific ironic node.

Change-Id: Id63d32501c97d2d27e5c61d8d46fff3e3c791505
Related-Bug: #1821764