Include all network devices in nova diagnostics
get_instance_diagnostics expected all interfaces
to have a <target> element with a "dev" attribute in
the instance XML. This is not the case for VFIO
interfaces (<interface type="hostdev">).
This caused an IndexError when looping over
the interfaces.

This patch fixes this issue by retrieving interfaces
data directly from the guest XML and adding nics
appropriately to the diagnostics object.

The new functional test has been left out of this
cherry-pick, since a lot of the test code that
supports the test is missing and would have to be
back-ported just for that one test, including a
ramification of other commit dependencies.
The functional code change itself is rather simple,
and not having this functional test present in
Rocky is considered to be low risk.

Change-Id: I8ef852d449e9e637d45e4ac92ffc5d1abd8d31c5
Closes-Bug: #1821798
(cherry picked from commit 1d4f64b190afc60b0c2a56de718209869c41cfb3)