Change the TODO to NOTE about instance multi-create
The _fill_provider_mapping() calls the placement to get traits of the
RPs in the allocation for each instance during build. In a multi create
request multiple instance might hit the same compute host so caching the
traits of the RPs on those compute host reduces the placement load.

However multi-create can be used with port's that has resource request
(those the one that triggers _fill_provider_mapping) because for multi
create to work the user needs to pass networks to the server create API
insted of pre-created Neutron ports. However ports with resource request
only supported if pre-created ports are passed to the server create API.

Therefore this patch corrects the TODO about multi-create in
_fill_provider_mapping.

Change-Id: I8e8248981b533cb4c37e89009f5839e41d9c6649