Add option to reserve all memory
If we don't reserve all memory, there's a swap file pre-created for
every VM on the ephemeral storage - even if it's unused.

But this option comes at the price of not being able to over-commit
memory.

Related issue: CCM-9589

Change-Id: I916f3ed8d32efc9de36b943546fee7440bbbf7ca
Note: This will later be reverted in favor of manual memory allocations.