Only call _fill_provider_mapping if claim succeeds
During re-schedule condutor takes the next Selecton object from the
host list and tries to allocate the requested resources on the host in
the Selection object. So far the conductor also tried to find the
resource provide mapping for such allocation even if the resource claim
is failed. This is unnecessary. This patch makes sure that mapping is
tried to be calculated if the claim succeeds first.

Change-Id: I9944398c38d11466d27c2a4b24035b26d264b000
Closes-Bug: #1822262