Mock time.sleep() in unit tests
A previous patch [1] attempted to make sure that tests would never
sleep() unless they really meant to. That turned out to be too invasive,
and too difficult to get right. But it revealed all of the places in the
nova codebase where real nonzero sleep()s were being done in test code.

Some of those are necessary: specifically the ones in functional test
code, which are doing things like _wait_for_server_status while other
threads are building (almost) real instances and such.

But almost all of the ones in unit tests are just wasting time. So this
commit mocks time.sleep() in those.

[1] If971c31380b1c5d1cc242846fb9c0779f03bbd19

Change-Id: I8ca348ab1a671b943c17dfec9ada2103aeb7cd84