Add placement as required project to functional py36 and 37
The functional tests are configured to install placement from git
so that they test with placement-from-master. In the py27 and py35
functional tests, placement is a part of required-projects, which
means that depends-on to a particular patch of placement can work
via tox-siblings[1].

This patch adds the same functionality for py36 and py37 functional
tests, which currently run only in the experimental queue.

[1] https://docs.openstack.org/infra/manual/zuulv3.html#installation-of-sibling-requirements

Change-Id: Ia40aba6d3036fea2bc62fdf76e8e221b552f1cdc