Don't report 'exiting' when mapping cells
When running 'nova-manage simple_cell_setup...' if there are not hosts
to map, but there remaining instances to map, an '..., exiting' message
is produced. This is misleading because "exiting" implies a return of
control to the user. That doesn't happen if there are many instances
left to inspect or map.

This change gets around that by getting rid of the exiting message
in the case where instance mapping can still happen.

Change-Id: I62b20a3676429b5cc756884275138566785b347e
Closes-Bug: #1821737