Leave brackets on Ceph IP addresses for libguestfs
Brackets are removed from all IPv6 addresses when getting Ceph monitor
addresses. This is correct for libvirt's XML files but not for
libguestfs.

This adds a 'strip_brackets' argument for get_mon_addrs method. It is
default 'True' to not interfere anywhere this is used with no arguments.

Uses strip_brackets=False when getting the hosts and ports for RBDImage,
which is used for libguestfs.

Change-Id: I378998a73046b9b9a33e30d0e06c4721d712da65
Closes-Bug: 1822689