Docs: emulator threads: clarify expected behavior
This patch adds a table of possible extra-spec/cpu_shared_set
combinations, and lists the expected behavior of instance emulator
threads for each case. This is an improvement over the previous doc,
which incorrectly implied that `share` was the default policy value,
and only covered what happens if cpu_shared_set is set.

Change-Id: Ia56607d0b70d5f46159b57ee1092ba0e37a02270