Remove MIN_COMPUTE_MULTIATTACH conditions in API
nova-compute service version 27 was added in Queens [1]
and the API started relying on it in Queens [2] when
creating a server with a multi-attach volume.

We are long past supporting computes older than Queens
so we can drop that minimum service version check code
from the server create API flow and related error handling.

Note that when attaching a multi-attach volume to a server
the version compatibility was handled in the compute RPC API
reserve_block_device_name method, but that code was removed
in Rocky [3]. However, the volume attach API still has to
handle the MultiattachSupportNotYetAvailable exception if
using cells v1 because it is raised from
ComputeCellsAPI._attach_volume.

[1] Ieec2bbd8a23c861c89bf9922517fa6a5562f1937
[2] I02120ef8767c3f9c9497bff67101e57e204ed6f4
[3] Ibcb6bf912b3fb69c8631665fef2832906ba338aa

Change-Id: I7700f66fd590b617333ae99c1b42986278b36263