Remove 'instance_info_cache_update_at_top'
This was being called within the 'InstanceInfoCache' object. Remove it
in its entirety, given that nothing should be calling it now.

An 'update_cells' argument is removed from a number of functions as
these all resulted in a call to the cells v1 RPC API, a call that is now
removed.

Part of blueprint remove-cells-v1

Change-Id: Ia8da8cd040d104d745bd7adf20438b8eec558680
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>