Do not log a warning about not using compute monitors
The compute_monitors configuration option is empty by
default, however, on nova-compute startup we get a warning
logged saying:

  Excluding nova.compute.monitors.cpu monitor virt_driver. \
  Not in the list of enabled monitors (CONF.compute_monitors).

This is unnecessary noise in the logs when nova-compute is
not configured to use monitors, so this change avoids the warning
log message in that case.

The compute_monitors configuration option help text is also
updated to (1) fix formatting and (2) remove the mention of the
numa_mem_bw monitor which never made it into nova (see
blueprint memory-bw).

Change-Id: I5bd204e0017f6f795e3cf13b34b75124ee23aa78
Closes-Bug: #1823207
(cherry picked from commit 8795f591d541227cd704019f19f3a78c74fc8ac7)