Handle PortLimitExceeded in POST /servers/{server_id}/os-interface
When attaching an interface to a server, if an existing port is not
specified, nova-compute will attempt to create a port on either the
user-specified network or the network that is available to the tenant.

If the tenant exceeds their port quota in neutron, a PortLimitExceeded
exception is raised up from nova-compute [1] which is not being handled
in the API controller code - which is fixed in this change.

Note that this is one of the few synchronous RPC call operations [2]
in the compute API so exceptions from the compute service will leak
back to the API like in this case and need to be handled to avoid a
500 response to the user.

The 403 response used here matches how PortLimitExceeded is handled in
the server create API [3].

[1] https://github.com/openstack/nova/blob/6ebb2c4ca/nova/network/neutronv2/api.py#L565
[2] https://github.com/openstack/nova/blob/6ebb2c4ca/nova/compute/rpcapi.py#L489
[3] https://github.com/openstack/nova/blob/6ebb2c4ca/nova/api/openstack/compute/servers.py#L688

Change-Id: I5687480a22542eea31d299442837bd64bf731285
Closes-Bug: #1823203