Fix SynchronousThreadPoolExecutorFixture mock of Future
The fixture is not mocking the Future.result() method correctly and an
attempt to access it results in the following error:

  AttributeError: Mock object has no attribute 'result'

Instead of mocking the submit() method, we can just swap out the
GreenThreadPoolExecutor for the SynchronousExecutor to make the
executor synchronous for testing.

A few changes were also needed in unit tests that are running with the
thread pool executor. Because the thread pool executor is created in
the compute manager's __init__ method, we need to patch the synchronous
executor in before calling the base test class's setUp method. Next, a
couple of unit tests were asserting that an exception would be raised
from the thread pool executor's submit() method but that will not
happen because if an exception is raised during a submitted task, that
exception will not be raised until Future.result() is called. And
Future.result() is never called in the compute manager.

Closes-Bug: #1823278

Change-Id: I719c222a5fa3d654e7397f126a1338e653b7bbcc