Add functional recreate test for bug 1823370
When evacuating a server in a multi-cell environment
we should be restricting the scheduling request during
evacuate to the cell in which the instance already exists
since we don't support cross-cell evacuate.

This adds a functional test to recreate the bug to show
that the scheduler is not restricted to the instance's
current cell when evacuating.

Change-Id: I56e20c84f25cc4961dc8d637c222b6f213c4d5f9
Related-Bug: #1823370