Restrict RequestSpec to cell when evacuating
When evacuating a server in a multi-cell environment
we need to restrict the scheduling request during
evacuate to the cell in which the instance already exists
since we don't support cross-cell evacuate.

This fixes the issue by restricting the RequestSpec to
the instance's current cell during evacuate in the same
way we do during unshelve.

Note that this should also improve performance when
rebuilding a server with a new image since we will only
look for the ComputeNode from the targeted cell rather
than iterate all enabled cells during scheduling.

Change-Id: I497180fb81fd966d1d3d4b54ac66d2609347583e
Closes-Bug: #1823370